+++ app/admin/merch/page.tsx
@@ -1,0 +1,5 @@
+import ProductManagementClient from "./MerchManagementClient";
+
+export default function MerchPage() {
+  return <ProductManagementClient />;
+}